[+] corrections linking question and video + rebase url routes api + update routes documentation /

+++ questions/views.py
@@ -4,23 +4,33 @@ from rest_framework.response import Response
 from .serializers import QuestionSerializer
 from .models import Question
 
-@api_view(['GET'])
-def getQuestions(request, video_id):
-    questions = Question.objects.filter(video_id=video_id)
-    if questions == []:
-        return Response({})
-    serializer = QuestionSerializer(questions, many=True)
-    return Response(serializer.data)
+@api_view(['POST'])
+def getQuestions(request):
+    data = request.data
+
+    questions = Question.objects.filter(video_id=data['video_id'])
+    if len(questions) > 0:
+        serializer = QuestionSerializer(questions, many=True)
+        return Response(serializer.data)
+    else: 
+        return Response("No comments attached to this video id")
 
 @api_view(['POST'])
-def createQuestion(request, video_id):
+def createQuestion(request):
     data = request.data
-    question = Question.objects.create(question=data['question'], author=data['author'], video_id=video_id)
-    serializer = QuestionSerializer(question, many=False)
-    return Response(serializer.data)
+    if data['question'] and data['author'] and data['video_id']:
+        question = Question.objects.create(question=data['question'] , author=data['author'], video_id=data['video_id'])
+        serializer = QuestionSerializer(question, many=False)
+        return Response(serializer.data)
+    else:
+        return Response("One of the fields is missing or corrupted")
 
 @api_view(['DELETE'])
-def deleteQuestion(request, video_id, id_question):
-    question = Question.objects.get(id=id_question)
-    question.delete()
-    return Response('Question was deleted')
+def deleteQuestion(request):
+    data = request.data
+    if data['id']:
+        question = Question.objects.get(id=data['id'])
+        question.delete()
+        return Response('Question was deleted or does not exists but the request is valid')
+    else:
+        return Response('Question id is unvalid or corrupted')
